Add fish configuration file

+++ install.py
@@ -8,9 +8,11 @@ import os
 import platform
 
 REPOSITORY = "https://github.com/charlax/dotfiles.git"
+INSTALL_SCRIPT = "https://raw.githubusercontent.com/charlax/dotvim/master/install.py"  # flake8: noqa
 DOTFILES_PATH = os.path.join(os.environ["HOME"], ".dotfiles")
 CONFIGURATION_FILES = (
     ("config/pudb/pudb.cfg", ".config/pudb/pudb.cfg"),
+    ("config/fish/config.fish", ".config/fish/config.fish"),
     "ctags/ctags",
     "ghci/ghci",
     "git/gitignore",
@@ -31,7 +33,7 @@ def force_symlink(src, dest):
     """Force symlink a file."""
     try:
         os.symlink(src, dest)
-    except OSError, e:
+    except OSError as e:
         if e.errno == errno.EEXIST:
             os.remove(dest)
             os.symlink(src, dest)
@@ -41,21 +43,21 @@ def create_folder(dest):
     """Create folder if not exists."""
     if os.path.exists(dest):
         return
-    print "Folder %s does not exist, creating" % dest
+    print("Folder %s does not exist, creating" % dest)
     os.makedirs(dest)
 
 
-def symlink_configuration_file(f, dest=None, force=False, base_path=None):
+def symlink_configuration_file(f,
+                               dest=None,
+                               force=False,
+                               base_path=None,
+                               dry_run=False):
     """Symlink a configuration to ~."""
-    if base_path:
-        if base_path == '.':
-            source_base_path = '.dotfiles'
-        else:
-            source_base_path = os.path.join(base_path, '.dotfiles')
-    else:
-        source_base_path = DOTFILES_PATH
-
     dest_base_path = base_path or os.environ["HOME"]
+
+    # Source is where the dotfiles are installed, it's always relative
+    # to where the files will be installed.
+    source_base_path = os.path.join(dest_base_path, '.dotfiles')
     source = os.path.join(source_base_path, f)
 
     if not dest:
@@ -63,31 +65,32 @@ def symlink_configuration_file(f, dest=None, force=False, base_path=None):
     else:
         dest = os.path.join(dest_base_path, dest)
 
-    print "Will link %s to %s" % (source, dest)
+    if dry_run:
+        print("Would symlink %s to %s" % (source, dest))
+        return
 
     create_folder(os.path.dirname(dest))
 
     if not force and (os.path.islink(dest) or os.path.exists(dest)):
-        print "Not symlinking '%s': already exists." % dest
+        print("Not symlinking '%s': already exists." % dest)
+        return
 
+    print("Symlinking '%s' -> '%s'." % (source, dest))
+    if not force:
+        try:
+            os.symlink(source, dest)
+        except OSError as e:
+            print("Could not symlink %s: %r" % (source, e))
     else:
-        print "Symlinking '%s' -> '%s'." % (source, dest)
-        if not force:
-            try:
-                os.symlink(source, dest)
-            except OSError as e:
-                print "Could not symlink %s: %r" % (source, e)
-        else:
-            force_symlink(source, dest)
-        logging.info("%s symlinked to %s" % (source, dest))
+        force_symlink(source, dest)
+    logging.info("%s symlinked to %s" % (source, dest))
 
 
 def clone_dotfile(repo, path):
     """Clone or update the dotfiles directory."""
-    if not os.path.exists(path):
-        os.system("git clone %s %s" % (repo, path))
-    else:
-        os.system("cd %s && git checkout master && git pull" % path)
+    if os.path.exists(path):
+        return
+    os.system("git clone %s %s" % (repo, path))
 
     if not os.path.exists(path):
         raise Exception("Dotfiles path '%s' does not exist" % path)
@@ -97,39 +100,50 @@ def symlink(args):
     """Symlink the files."""
     for f in CONFIGURATION_FILES:
         if isinstance(f, (tuple, list)):
-            symlink_configuration_file(*f,
-                                       base_path=args.symlink_base,
-                                       force=args.symlink_force)
+            symlink_configuration_file(
+                *f,
+                base_path=args.symlink_base,
+                force=args.symlink_force,
+                dry_run=args.dry_run)
         else:
-            symlink_configuration_file(f,
-                                       base_path=args.symlink_base,
-                                       force=args.symlink_force)
+            symlink_configuration_file(
+                f,
+                base_path=args.symlink_base,
+                force=args.symlink_force,
+                dry_run=args.dry_run)
 
 
 def main(args):
     """Install the dotfiles."""
-    # Must install before symlinking. Otherwise, some directory would
-    # not exist, in particular ~/.virtualenvs
     clone_dotfile(REPOSITORY, DOTFILES_PATH)
     symlink(args)
 
     if args.with_dotvim:
-        print "Installing dotvim..."
-        os.system("curl https://raw.githubusercontent.com/charlax/dotvim/master/install.py -o install_dotvim.py")  # noqa
+        print("Installing dotvim...")
+        os.system("curl %s  -o install_dotvim.py" % INSTALL_SCRIPT)
         os.system("python install_dotvim.py")
         os.remove("install_dotvim.py")
 
-    print "Install complete."
+    print("Install complete.")
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Install charlax's dotfiles.")
-    parser.add_argument("--with-dotvim", action="store_true",
-                        help="Also install charlax's dotvim repository")
-    parser.add_argument("--symlink-force", "-f", action="store_true",
-                        help="Force symlink the files")
-    parser.add_argument("--symlink-base",
-                        help="Base path for symlinks")
+    parser.add_argument(
+        "--with-dotvim",
+        action="store_true",
+        help="Also install charlax's dotvim repository")
+    parser.add_argument(
+        "--symlink-force",
+        "-f",
+        action="store_true",
+        help="Force symlink the files")
+    parser.add_argument("--dry-run", action="store_true", help="Dry run")
+    parser.add_argument(
+        "--symlink-base",
+        help="Base path for symlinks, defaults to ~",
+    )
+
     args = parser.parse_args()
 
     main(args)
